arithmetic: Split `bn_sqr8x_mont` into its own submodule.
```
git difftool HEAD^1:src/arithmetic/ffi.rs \
                    src/arithmetic/x86_64_mont.rs
```
