+++ src/arithmetic/ffi.rs
@@ -95,55 +95,3 @@ pub(super) unsafe fn bn_mul_mont_ffi<Cpu, const LEN_MIN: usize, const LEN_MOD: u
         })
         .map_err(LimbSliceError::len_mismatch)
 }
-
-#[cfg(target_arch = "x86_64")]
-#[inline]
-pub(super) fn bn_sqr8x_mont(
-    in_out: &mut [Limb],
-    n: &[Limb],
-    n0: &N0,
-    cpu: crate::cpu::Features,
-) -> core::ops::ControlFlow<(), ()> {
-    use crate::cpu::{
-        intel::{Adx, Bmi2},
-        GetFeature as _,
-    };
-    use core::ops::ControlFlow;
-
-    prefixed_extern! {
-        // `r` and/or 'a' may alias.
-        // XXX: BoringSSL declares this to return `int`.
-        fn bn_sqr8x_mont(
-            rp: *mut Limb,
-            ap: *const Limb,
-            mulx_adx_capable: Limb,
-            np: *const Limb,
-            n0: &N0,
-            num: c::size_t);
-    }
-
-    if n.len() < 8 {
-        return ControlFlow::Continue(());
-    }
-    if n.len() % 8 != 0 {
-        return ControlFlow::Continue(());
-    }
-    // Avoid stack overflow from the alloca inside.
-    if n.len() > MAX_LIMBS {
-        return ControlFlow::Continue(());
-    }
-
-    let rp = in_out.as_mut_ptr();
-    let ap = in_out.as_ptr();
-    let mulx_adx: Option<(Adx, Bmi2)> = cpu.get_feature();
-    // `Limb::from(mulx_adx.is_some())`, but intentionally branchy.
-    let mulx_adx_capable = match mulx_adx {
-        Some(_) => Limb::from(true),
-        None => Limb::from(false),
-    };
-    let np = n.as_ptr();
-    let num = n.len();
-    unsafe { bn_sqr8x_mont(rp, ap, mulx_adx_capable, np, n0, num) };
-
-    ControlFlow::Break(())
-}
